GetCustomerById changed

+++ dotNet5782_2344_7364/BL/DroneToList.cs
@@ -13,7 +13,7 @@ namespace IBL
             public int Id { set; get; }
             public string Model { set; get; }
             public Enums.WeightCategories Weight { set; get; }
-            public float Battery { set; get; }
+            public double Battery { set; get; }
             public Enums.DroneStatuses DroneStatuses { set; get; }
             public Location CurrentLocation { set; get; }
             public int NumberOfParcelInTransit { set; get; }
@@ -21,7 +21,7 @@ namespace IBL
             {
                 return $" Drone #{Id}: \n Battery = {Battery} Weight = {Weight} \n Battery = {Battery} \n Drone Status = {DroneStatuses} \n Current location = {CurrentLocation} \n Number of parcel in transit = {NumberOfParcelInTransit} \n";
             }
-            public DroneToList(int id, string model, Enums.WeightCategories weight , float battery, Enums.DroneStatuses droneStatuses, Location location, int numberOfParcels)
+            public DroneToList(int id, string model, Enums.WeightCategories weight , double battery, Enums.DroneStatuses droneStatuses, Location location, int numberOfParcels)
             {
                 Id = id;
                 Model = model;
